Lesson 6. HW6. Main task. Optional task partially.
Realized all subtasks of Main task, but app doesn't work properly -
doesn't open user's meals page because of java.lang.NullPointerException.
In the Optional task realized 1, 2, 3 and partially 5 subtask.
In the 5 subtask DataJpaUserServiceTest not passed getWithMeals;
in the JdbcUserServiceTest not passed create, getAll, update, get, getByEmail
methods because of lack of 4 subtask realization.